+++ apps/api-service/src/products/wealthanalyzer/asset/asset.service.ts
@@ -99,7 +99,6 @@ export class AssetService {
     }
   }
 
-  @OnEvent(EventMap.GetAssetDetailsById)
   async findAssetById(reqUserId: string, assetId: string) {
     try {
       const asset = await this.queryBus.execute<FindAssetByIdQuery, Asset>(