+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Rebranding of FinanceNews app to Discover
 - Added turn off feature for Intelligence & data usage agreement
+- Revamped architecture of intelligence summarizer
 - Improve analytics code to have generic function that adds analytics to entities
 - Introduce ExpenseTrack app
 - Improve authentication with approval codes & better email template